PoC/requirements.txt
pytest
pytest-cov
pylint
pymongo
kafka-python
mongoengine
flask